X3. Build and Display the Watchlist UI #47 feat: add SVGR plugin for SVG imports and update package.json
style: adjust line clamp in App.css for text truncation

refactor: separate Main and WatchlistView components from App.tsx

feat: implement watchlist functionality in MovieCard component

feat: add logo SVG asset for branding in the header

fix: include Bootstrap JS in main.tsx for proper navbar functionality

feat: create WatchlistView component to display saved movies
