+++ Movie-Project/src/App.css
@@ -1,8 +1,8 @@
 .truncate {
 	display: -webkit-box;
-	line-clamp: 3;
+	line-clamp: 2;
 	/* number of lines */
-	-webkit-line-clamp: 3;
+	-webkit-line-clamp: 2;
 	-webkit-box-orient: vertical;
 	overflow: hidden;
 }
